+++ xyzpy/utils.py
@@ -124,6 +124,37 @@ def progbar(it=None, nb=False, **kwargs):
     return tqdm.tqdm(it, **settings)
 
 
+def getsizeof(obj):
+    """Compute the real size of a python object. Taken from
+
+    https://stackoverflow.com/a/30316760/5640201
+    """
+    import sys
+    from types import ModuleType, FunctionType
+    from gc import get_referents
+
+    # Custom objects know their class.
+    # Function objects seem to know way too much, including modules.
+    # Exclude modules as well.
+    excluded = type, ModuleType, FunctionType
+    if isinstance(obj, excluded):
+        raise TypeError(
+            'getsize() does not take argument of type: {}'.format(type(obj)))
+
+    seen_ids = set()
+    size = 0
+    objects = [obj]
+    while objects:
+        need_referents = []
+        for obj in objects:
+            if not isinstance(obj, excluded) and id(obj) not in seen_ids:
+                seen_ids.add(id(obj))
+                size += sys.getsizeof(obj)
+                need_referents.append(obj)
+        objects = get_referents(*need_referents)
+    return size
+
+
 class Timer:
     """A very simple context manager class for timing blocks.